+++ packages/eyes-api/test/unit/Runner.spec.ts
@@ -1,24 +1,30 @@
-const assert = require('assert')
-const makeSDK = require('../utils/fake-sdk')
-const api = require('../../dist')
+import type * as types from '@applitools/types'
+import assert from 'assert/strict'
+import * as api from '../../src'
 
-describe('Eyes', () => {
-  class Eyes extends api.Eyes {}
+const makeSDK = require('../utils/fake-sdk')
+
+describe('Runner', () => {
+  let sdk: types.Core<any, any, any> & {history: Record<string, any>[]; settings: Record<string, any>}
+  const driver = {isDriver: true}
+
+  class Eyes extends api.Eyes {
+    protected static get _spec() {
+      return sdk
+    }
+  }
   class ClassicRunner extends api.ClassicRunner {}
   class VisualGridRunner extends api.VisualGridRunner {}
-  const driver = {isDriver: true}
-  let sdk
 
   beforeEach(() => {
     sdk = makeSDK()
-    Eyes._spec = sdk
   })
 
   it('should return empty test summary from "getAllTestResults" method if no eyes instances were attached', async () => {
     const runner = new ClassicRunner()
     const summary = await runner.getAllTestResults()
     assert.ok(summary instanceof api.TestResultsSummary)
-    assert.strictEqual(Array.from(summary).length, 0)
+    assert.equal(Array.from(summary).length, 0)
   })
 
   it('should return empty test summary from "getAllTestResults" method if only attached eyes instance is disabled', async () => {
@@ -29,6 +35,6 @@ describe('Eyes', () => {
     eyes.close(false)
     const summary = await runner.getAllTestResults()
     assert.ok(summary instanceof api.TestResultsSummary)
-    assert.strictEqual(Array.from(summary).length, 0)
+    assert.equal(Array.from(summary).length, 0)
   })
 })